# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clusters Metadata API",
    "description": "Canonical API for managed-cluster metadata.",
    "version": "development"
  },
  "paths": {
...
    "/v3/features": {
      "get": {
        "tags": [
          "features"
        ],
        "summary": "List Features",
        "operationId": "list_features_v3_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "features"
        ],
        "summary": "Create Feature",
        "operationId": "create_feature_v3_features_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/v3/features/{feature_id}": {
      "delete": {
        "tags": [
          "features"
        ],
        "summary": "Delete Feature",
        "operationId": "delete_feature_v3_features__feature_id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "feature_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feature Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v3/locks": {
      "get": {
        "tags": [
          "locks"
        ],
        "summary": "List Locks",
        "operationId": "list_locks_v3_locks_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster Name"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "locked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Locked"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterLockListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v3/clusters/{cluster_name}/lock": {
      "post": {
        "tags": [
          "locks"
        ],
        "summary": "Acquire Lock",
        "operationId": "acquire_lock_v3_clusters__cluster_name__lock_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquireClusterLockRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquireClusterLockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "locks"
        ],
        "summary": "Release Lock",
        "operationId": "release_lock_v3_clusters__cluster_name__lock_delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseClusterLockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseClusterLockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v3/operations": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "List Operations",
        "operationId": "list_operations_v3_operations_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "operation_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation Type"
            }
          },
          {
            "name": "operation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation Status"
            }
          },
          {
            "name": "cicd_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uri",
                  "maxLength": 2083,
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cicd Url"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v3/clusters/{cluster_name}/operations": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "List Cluster Operations",
        "operationId": "list_cluster_operations_v3_clusters__cluster_name__operations_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Create Operation",
        "operationId": "create_operation_v3_clusters__cluster_name__operations_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcquireClusterLockRequest": {
        "properties": {
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "timeout_minutes": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 0.0,
            "title": "Timeout Minutes",
            "default": 360
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AcquireClusterLockRequest"
      },
      "AcquireClusterLockResponse": {
        "properties": {
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "timeout_at": {
            "type": "string",
            "format": "date-time",
            "title": "Timeout At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cluster_name",
          "message",
          "token",
          "timeout_at"
        ],
        "title": "AcquireClusterLockResponse"
      },
...
      "ClusterLockListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClusterLockRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "ClusterLockListResponse"
      },
      "ClusterLockRead": {
        "properties": {
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "locked": {
            "type": "boolean",
            "title": "Locked"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "timeout_at": {
            "type": "string",
            "format": "date-time",
            "title": "Timeout At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cluster_name",
          "locked",
          "timeout_at",
          "created_at",
          "updated_at"
        ],
        "title": "ClusterLockRead"
      },
...
      "FeatureConstraint": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "operator": {
            "$ref": "#/components/schemas/FeatureOperator"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "operator"
        ],
        "title": "FeatureConstraint"
      },
      "FeatureCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feature_type": {
            "$ref": "#/components/schemas/FeatureType",
            "default": "optional"
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FeatureConstraint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "namespaced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespaced"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FeatureCreate",
        "description": "Schema for creating a feature (used inside release creation)."
      },
      "FeatureDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "FeatureDeleteResponse"
      },
      "FeatureListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeatureRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "FeatureListResponse"
      },
      "FeatureOperator": {
        "type": "string",
        "enum": [
          "equals",
          "not_equals",
          "in",
          "not_in"
        ],
        "title": "FeatureOperator"
      },
      "FeatureRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feature_type": {
            "$ref": "#/components/schemas/FeatureType",
            "default": "optional"
          },
          "dependencies": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependencies"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FeatureConstraint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "namespaced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespaced"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "FeatureRead",
        "description": "Feature as returned from the API."
      },
...
      "FeatureType": {
        "type": "string",
        "enum": [
          "core",
          "optional"
        ],
        "title": "FeatureType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationCreate": {
        "properties": {
          "operation_type": {
            "type": "string",
            "title": "Operation Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cicd_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Cicd Url"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_type",
          "status",
          "cicd_url"
        ],
        "title": "OperationCreate"
      },
      "OperationListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count",
          "items"
        ],
        "title": "OperationListResponse"
      },
      "OperationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "operation_type": {
            "type": "string",
            "title": "Operation Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cicd_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Cicd Url"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "operation_type",
          "status",
          "cicd_url"
        ],
        "title": "OperationRead"
      },
...
      "ReleaseClusterLockRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token"
        ],
        "title": "ReleaseClusterLockRequest"
      },
      "ReleaseClusterLockResponse": {
        "properties": {
          "cluster_name": {
            "type": "string",
            "title": "Cluster Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cluster_name",
          "message"
        ],
        "title": "ReleaseClusterLockResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}